Move Homepage from App.js to pages
