+++ src/pages/Rarity.jsx
@@ -1,0 +1,10 @@
+import React from "react"
+
+
+function Rarity() {
+  return (
+    <h1>Hello Rarity</h1>
+  )
+}
+
+export default Rarity
